finalizando proyecto, solo falta arreglar bugs en lista de Ofertas

+++ application/views/oferta/index.php
@@ -1,5 +1,5 @@
 <!-- main-heading -->
-<h2 class="main-title-w3layouts mb-2 text-center">Facultades</h2>
+<h2 class="main-title-w3layouts mb-2 text-center">Ofertas</h2>
 <!--// main-heading -->
 <?php if (isset($response)) { ?>
 	<div class="col-md-6 offset-md-3 space-bot-md">
@@ -34,12 +34,31 @@
 	                    <td><?php echo $oferta['ofer_destino'] ?></td>	
 	                    <td><?php echo $oferta['ofer_destino_dep'] ?></td>
 	                    <td><?php echo $oferta['ofer_descripcion'] ?></td>
-	                    <td>
-	                    	<div class="btn-group">
-	                    		<a href="<?php echo base_url(); ?>index.php/Oferta/edit/<?php echo $oferta['ofer_id'];?>" class="btn btn-sm btn-warning">Editar</a>
-	                    		<a href="<?php echo base_url(); ?>index.php/Oferta/confirm_delete/<?php echo $oferta['ofer_id'];?>" class="btn btn-sm btn-danger">Eliminar</a>
-	                    	</div>
-	                    </td>
+	                    <?php 
+	                    $rep = 0;
+	                    foreach ($llamados as $key => $llamado) {
+	                    	if($llamado['ofer_id'] == $oferta['ofer_id'])
+	                    	{
+	                    		$rep=1;
+	                    		if($llamado['llam_status']=='finalizado'){
+			                    	echo "<td>Oferta Finalizada</td>";
+			                    }else if($llamado['llam_status']=='desierto'){
+			                    	echo "<td>Oferta Desierta</td>";
+			                   	}			                    
+			                    break;	                     
+			                }else if($rep==0) {?>
+								<td>	                    	
+			                    	<div class="btn-group">
+			                    		<a href="<?php echo base_url(); ?>index.php/Llamado/create/<?php echo $oferta['ofer_id'];?>" class="btn btn-sm btn-success">Generar Llamado</a>
+			                    		<a href="<?php echo base_url(); ?>index.php/Oferta/curriculum/<?php echo $oferta['ofer_id'];?>" class="btn btn-sm btn-primary">Ver Currículums</a>
+			                    		<a href="<?php echo base_url(); ?>index.php/Oferta/edit/<?php echo $oferta['ofer_id'];?>" class="btn btn-sm btn-warning">Editar</a>
+			                    		<a href="<?php echo base_url(); ?>index.php/Oferta/confirm_delete/<?php echo $oferta['ofer_id'];?>" class="btn btn-sm btn-danger">Eliminar</a>
+				                    	</div>
+				                    </td>
+			                    <?php
+			                    break;
+			                	}
+	                    }?>	                    
 	                </tr>
 	            <?php endforeach ?>
 	        </tbody>
